src/requirements.txt
pandas
python-dotenv
requests
openpyxl
flask
flask-cors
wget
ibm_watson_machine_learning
ibm_watsonx_ai
ibm_cloud_sdk_core
pyarrow


docling
langchain
langchain-ibm
langchain-core
ibm-watsonx-ai

crewai
crewai-tools

duckduckgo-search

ibmpairs
rasterio
folium
configparser
matplotlib